Add component for post view, insert in view message and reply to msg
