+++ src/shared/features/GrandparentViews/GrandparentReply/GrandparentReplyOpts.tsx
@@ -1,11 +1,12 @@
 import React, { useState } from 'react';
 import { useHistory } from "react-router-dom";
 
-import { Box } from '@material-ui/core';
+import { GrandparentPostView } from "../GrandparentPostView";
 import { Post } from "../../../models/post.model";
 import { replyOptionIcons } from "../../../../Icons";
 import GetEmojiReply from "./components/GetEmojiReply";
 import GrandparentLayout from "../GrandparentLayout";
+
 
 interface IGrandparentReplyOpts {
   post: Post;
@@ -14,7 +15,6 @@ interface IGrandparentReplyOpts {
 const GrandparentReplyOpts: React.FC<IGrandparentReplyOpts> = ({post}) => {
 
   const [EmojiReplyOpen, setEmojiReplyOpen] = useState<boolean>(false);
-  const boxContent = post.message + "\n\n" + post.photoURL;
   let history = useHistory();
 
   return (
@@ -22,7 +22,7 @@ const GrandparentReplyOpts: React.FC<IGrandparentReplyOpts> = ({post}) => {
 
         <GrandparentLayout
           headerText={"Reply to Letter from "}
-          boxContent={boxContent}
+          boxContent={GrandparentPostView}
           buttonText={["Return to Messages", "Smiley", "Voice Message", "Your Picture"]}
           buttonActions={[
             () => history.push("/posts"),
@@ -33,19 +33,6 @@ const GrandparentReplyOpts: React.FC<IGrandparentReplyOpts> = ({post}) => {
           buttonIcons={[replyOptionIcons.closedEnvelope, replyOptionIcons.emoji,
                         replyOptionIcons.voicemail, replyOptionIcons.photo]}
           />
-
-         <Box className="todo">
-            <h3>To do items:</h3>
-            <ul>
-              <li>Fix Box size</li>
-              <li>Create replies:</li>
-                <ul>Send Emoji</ul>
-                <ul>Send "Voice Message"</ul>
-                <ul>Send Selfie</ul>
-              <li>"Sent reply" View</li>
-              <li>Return to Post (requires return to Inbox + open modal</li>
-            </ul>
-        </Box>
 
         <GetEmojiReply
           isOpen={EmojiReplyOpen}
